rekelvin: fix S3 connection credentials (#808)
I wasn't explicitly passing in the S3 credentials, which was fine on my
local machine, since boto looked for credentials in environment files.
But of course this broke on production.

Explicitly get the credentials from the settings and use.

Also changed the name of the bucket variable to aws_bucket to be
consistent with the new var names.