change the download buttons for files on cm page to include the file extension
also: moved the logic for getting an appropriate icon from views.py to
models.py - hoped to call it from the course materials page, but had
trouble getting certain items to look good in that layout, so skipped
it and left the default file icon there.  And: got rid of the extra 2
pixels I had added to the video download buttons to make them line up
with the file download button that used to say "slides".