Parameterize init script, fix startup bug (#1371)
This fixes an insidious bug where we were using the name of the command
itself to figure out where to get our defaults from.  This worked fine
when calling the init script directly, but not when calling thorugh a
symlink, which is the case with rc2.d/S20... for example.