Interactive: examrecord a running score, not one-time (#1449)
Instead of setting the score to the last score, adjust the examrecord
score by the current value.  This could be the wrong behavior if someone
answers the same question wrong twice with the same wrong answer, for
example.  Should they be penalized twice?  Using this scorer, they will.

Also added some docstrings and formatting cleanups while I was in there.