videocheck: separate term and class options (fix #1786)
Now have two options, -c for the class ("db") and -t for the term
("Winter2013").  Useful now that we have multiple terms, and I want to
look only at videos for the current term.

Although this changes the meaning of the "-c" option from a handle
("db--Winter2013") to a class, I'm the only one that uses this utility
and it's not in any scripts that I know about.  And this will be more
useful this way.